CHORE: create Comment directory and move Comment.js
- change import statement in Comments.js
